Uploaded Patient List backend complete.
The list shows all the patient data uploaded by the current viewer, as required in the specs. This is done by matching current user ID with the doctor ID associated with each patient profile.

The list shows only 7 entries. Each entry, upon clicking, should ideally open up another activity that displays the entire 15 entries in a neat column. Work is to be commenced on that.