Removed Auth with Facebook and Twitter to even further simplify login flow.
Everyone that uses the app is going to have a Google account, since they'll need it to download the app from the Play Store, so it logically followed that to eliminate clutter, Google should be the only Auth Provider.

Gradle Version has also been updated in the app, bringing slight increases in speed in some components.